% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{assert_jster}
\alias{assert_jster}
\title{Assert jster app success}
\usage{
assert_jster(dt)
}
\arguments{
\item{dt}{\code{\link[=data.frame]{data.frame()}} of information returned from \code{\link[=run_jster]{run_jster()}} or \code{\link[=run_jster_apps]{run_jster_apps()}}}
}
\description{
Verify that all jster apps tested successfully
}
