% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{js_script}
\alias{js_script}
\title{JavaScript helper}
\usage{
js_script(...)
}
\arguments{
\item{...}{JavaScript text to be put in a script.}
}
\description{
Wraps supplied text in an \code{htmltools::tags$script} call after turning it into \code{JS} code.
}
