% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_headless.R
\name{run_headless_mac}
\alias{run_headless_mac}
\alias{run_headless_linux}
\title{Run many applications with shinyjster enabled}
\usage{
run_headless_mac(
  apps = apps_to_test(),
  port = 8000,
  host = "127.0.0.1",
  debug_port = 9222
)

run_headless_linux(
  apps = apps_to_test(),
  port = 8000,
  host = "127.0.0.1",
  debug_port = 9222
)
}
\arguments{
\item{apps}{Vector of \code{appDir} values}

\item{port}{The TCP port that the application should listen on. If the
\code{port} is not specified, and the \code{shiny.port} option is set (with
\code{options(shiny.port = XX)}), then that port will be used. Otherwise,
use a random port.}

\item{host}{The IPv4 address that the application should listen on. Defaults
to the \code{shiny.host} option, if set, or \code{"127.0.0.1"} if not. See
Details.}

\item{debug_port}{Port for chrome debugger}
}
\description{
Run many applications with shinyjster enabled
}
