% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selenium.R
\name{selenium_chrome}
\alias{selenium_chrome}
\alias{selenium_firefox}
\alias{selenium_edge}
\alias{selenium_ie}
\title{Selenium browsers}
\usage{
selenium_chrome(timeout = 2 * 60, headless = FALSE)

selenium_firefox(timeout = 2 * 60, headless = FALSE)

selenium_edge(timeout = 2 * 60, headless = FALSE)

selenium_ie(timeout = 2 * 60, headless = FALSE)
}
\arguments{
\item{timeout}{Number of seconds before selenium closes the browser}

\item{headless}{Logical which determines if the browser can run headless}
}
\description{
Opens a selenium driven browser and waits until shinyjster is finished.
}
\details{
This function assumes selenium is installed and all appropriate web browsers are installed.
}
\section{Functions}{
\itemize{
\item \code{selenium_chrome}: Opens a Chrome web browser

\item \code{selenium_firefox}: Opens a Firefox web browser

\item \code{selenium_edge}: Opens an Edge web browser

\item \code{selenium_ie}: Opens an IE web browser
}}

