% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{shinyjster_js_dependencies}
\alias{shinyjster_js_dependencies}
\title{shinyjster HTML Dependencies}
\usage{
shinyjster_js_dependencies()
}
\value{
\link[htmltools:htmlDependency]{htmltools::htmlDependency}'s to allow shinyjster to function.
}
\description{
shinyjster HTML Dependencies
}
