% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{shinyjster_server}
\alias{shinyjster_server}
\title{Shiny server helper}
\usage{
shinyjster_server(input, output, session)
}
\arguments{
\item{input, output, session}{Shiny server function parameters}
}
\description{
Function to be called within the shiny server
}
