% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{shinyjster_ui}
\alias{shinyjster_ui}
\title{Shiny UI helper}
\usage{
shinyjster_ui()
}
\description{
Function to be called first inside the definition of the Shiny UI.
}
\details{
This function will add the shinyjster JS dependencies and add a text based progress bar in the bottom left corner of the application.
}
