% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testthat_shinyjster.R
\name{testthat_shinyjster}
\alias{testthat_shinyjster}
\title{Test shinyjster on all browsers for shinycoreci}
\usage{
testthat_shinyjster(
  test_name = NULL,
  app_dir = "../../",
  ...,
  browsers = c("chrome", "firefox", "edge"),
  timeout = 2 * 60,
  dimensions = "1200x1200"
)
}
\arguments{
\item{test_name}{suffix to add to the test name for \code{testthat::test_that(NAME, {})}}

\item{app_dir}{Defaults the app in the directory above}

\item{...}{Ignored}

\item{browsers}{Names of each browser to be tested.}

\item{timeout, dimensions}{Parameters to be supplied to each browser}
}
\description{
This method will test a single application on all of the available browsers
shinyjster can test with on the given platform. For each browser, the app
will be tested using \code{\link[=test_jster]{test_jster()}}.
}
\details{
This method should be called from a test file in the \verb{./tests/testthat/} directory.

For each browser, a new \code{testthat::test_that()} test will be run. This allows
for browsers to not effect the other tests.
}
\seealso{
\code{\link[=test_jster]{test_jster()}}
}
