% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_jster.R
\name{use_jster}
\alias{use_jster}
\title{Create Shinyjster test file}
\usage{
use_jster(appDir = ".")
}
\arguments{
\item{appDir}{Location of shiny application to test}
}
\description{
This creates a testing file to be used with \code{shiny::runTests('.')}.  It will call \code{\link[=test_jster]{test_jster()}} which will cycle through all avaiable shinyjster selenium browsers.
}
\seealso{
\code{\link[=test_jster]{test_jster()}}
}
