% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_headless.R
\name{run_headless}
\alias{run_headless}
\title{Run many applications with shinyjster enabled}
\usage{
run_headless(
  apps = apps_to_test(),
  port = 8000,
  host = "127.0.0.1",
  debug_port = NULL,
  browser = c("chrome", "firefox"),
  type = c("lapply", "parallel", "callr"),
  assert = TRUE
)
}
\arguments{
\item{apps}{Vector of \code{appDir} values}

\item{port}{The TCP port that the application should listen on. If the
\code{port} is not specified, and the \code{shiny.port} option is set (with
\code{options(shiny.port = XX)}), then that port will be used. Otherwise,
use a random port.}

\item{host}{The IPv4 address that the application should listen on. Defaults
to the \code{shiny.host} option, if set, or \code{"127.0.0.1"} if not. See
Details.}

\item{debug_port}{Port for chrome debugger}

\item{browser}{Browser to be set for testing.  If using \code{"chrome"} or \code{"firefox"}, the system application will be used.}

\item{type}{Single value to determine how applications are executed. \describe{
\item{\code{'parallel'}}{Runs apps using \code{parallel::mclapply} using \code{cores} cores}
\item{\code{'callr'}}{Runs apps using \code{callr::r_bg} using \code{cores} cores}
\item{\code{'serial'}}{Runs apps one after another using \code{lapply}. \code{port} is only used with option \code{'serial'}}
}}

\item{assert}{Logical which determines if \link{assert_jster} should be run on the final output}
}
\description{
Run many applications with shinyjster enabled
}
