% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run_jster}
\alias{run_jster}
\alias{run_jster_apps}
\title{Run an application with shinyjster enabled}
\usage{
run_jster(appDir, port = 8000, host = "127.0.0.1")

run_jster_apps(
  apps,
  type = c("parallel", "callr", "serial"),
  cores = parallel::detectCores(),
  port = 8000,
  host = "127.0.0.1"
)
}
\arguments{
\item{appDir}{The application to run. Should be one of the following:
\itemize{
\item A directory containing \code{server.R}, plus, either \code{ui.R} or
a \code{www} directory that contains the file \code{index.html}.
\item A directory containing \code{app.R}.
\item An \code{.R} file containing a Shiny application, ending with an
expression that produces a Shiny app object.
\item A list with \code{ui} and \code{server} components.
\item A Shiny app object created by \code{\link[shiny:shinyApp]{shinyApp()}}.
}}

\item{port}{The TCP port that the application should listen on. If the
\code{port} is not specified, and the \code{shiny.port} option is set (with
\code{options(shiny.port = XX)}), then that port will be used. Otherwise,
use a random port.}

\item{host}{The IPv4 address that the application should listen on. Defaults
to the \code{shiny.host} option, if set, or \code{"127.0.0.1"} if not. See
Details.}

\item{apps}{Vector of \code{appDir} values}

\item{type}{Single value to determine how applications are executed. \describe{
\item{\code{'parallel'}}{Runs apps using \code{parallel::mclapply} using \code{cores} cores}
\item{\code{'callr'}}{Runs apps using \code{callr::r_bg} using \code{cores} cores}
\item{\code{'serial'}}{Runs apps one after another using \code{lapply}. \code{port} is only used with option \code{'serial'}}
}}

\item{cores}{Number of cores (if needed) to execute on.}
}
\description{
Run an application with shinyjster enabled
}
