% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selenium.R
\name{selenium_chrome}
\alias{selenium_chrome}
\alias{selenium_firefox}
\alias{selenium_edge}
\alias{selenium_ie}
\title{Selenium browsers}
\usage{
selenium_chrome(
  timeout = 2 * 60,
  dimensions = "1200x1200",
  headless = !interactive(),
  verbose = TRUE
)

selenium_firefox(
  timeout = 2 * 60,
  dimensions = "1200x1200",
  headless = !interactive(),
  verbose = TRUE
)

selenium_edge(timeout = 2 * 60, dimensions = "1200x1200", verbose = TRUE)

selenium_ie(timeout = 2 * 60, dimensions = "1200x1200", verbose = TRUE)
}
\arguments{
\item{timeout}{Number of seconds before selenium closes the browser}

\item{dimensions}{A string in the form of \verb{"WIDTHxHEIGHT"}. Ex: \code{"1800x1200"}}

\item{headless}{Logical which determines if the browser can run headless. Defaults to \code{TRUE} where possible.}

\item{verbose}{Logical which determines if the selenium output is displayed as it's received}
}
\description{
Opens a selenium driven browser and waits until shinyjster is finished.
}
\details{
This function assumes selenium is installed and all appropriate web browsers are installed.
}
\section{Functions}{
\itemize{
\item \code{selenium_chrome}: Opens a Chrome web browser

\item \code{selenium_firefox}: Opens a Firefox web browser

\item \code{selenium_edge}: Opens an Edge web browser

\item \code{selenium_ie}: Opens an IE web browser
}}

