% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_jster.R
\name{test_jster}
\alias{test_jster}
\title{Test shinyjster app on all browsers}
\usage{
test_jster(
  browsers = c(selenium_chrome(), selenium_firefox(), if (platform() == "win")
    c(selenium_edge(), selenium_ie())),
  appDir = "../",
  assert = TRUE,
  host = "127.0.0.1",
  port = NULL
)
}
\arguments{
\item{browsers}{By default, as many browsers as selenium support on the given platform}

\item{appDir}{Defaults the app in the directory above}

\item{assert}{A logical value that determines if the result should be validate if the return value passes all tests}

\item{host, port}{Used when running the shiny application}
}
\value{
A data frame with the columns \code{appDir}, \code{successful}, \code{returnValue}, and \code{browser}. One row of information per \code{browser}.
}
\description{
This method will test your shiny application using the shinyjster code you have provided on all of the available browsers shinyjster can test with on the given platform.
Each app will be ran in a separate R session using \code{run_jster_apps(type = 'serial')}.
}
\seealso{
\code{\link[=run_jster]{run_jster()}}, \code{\link[=use_jster]{use_jster()}}
}
